+++ packages/vue-styleguidist/src/types/Example.d.ts
@@ -1,14 +1,13 @@
 /* eslint-disable import/prefer-default-export */
 
-interface CompilableContent {
-	raw: string
-	compiled?: string
-}
-
 export interface Example {
 	type: 'code' | 'text'
 	lang: string
-	content: string | CompilableContent
+	content: { raw: string; compiled: string } | string
 	settings: { [key: string]: string }
 	compiled?: string
+}
+
+export interface ExampleLoader extends Example {
+	content: string
 }